pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "semantic_indexer"
version = "0.1.0"
description = "Semantic Document Indexer"
dependencies = [
    "python-dotenv==1.1.0",
    "psycopg2-binary==2.9.11",
    "google-genai==1.60.0",
    "pypdf==6.6.0",
    "python-docx==1.2.0",
    "nltk==3.9.2"
]
requires-python = ">=3.12"

[project.optional-dependencies]
test = [
    "pytest==8.0.0"
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools]
package-dir = {"" = "src"}
